Remove evaluator and merge template with model
